+++ AOS_BE/src/main/java/com/aos/AOSBE/Service/ProductItemsService.java
@@ -1,12 +1,14 @@
 package com.aos.AOSBE.Service;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
+import java.util.*;
 import java.util.stream.Collectors;
 
+import com.aos.AOSBE.Entity.ProductImages;
+import com.aos.AOSBE.Repository.*;
+import jakarta.annotation.PostConstruct;
+import org.springframework.ai.document.Document;
+import org.springframework.ai.vectorstore.qdrant.QdrantVectorStore;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -24,14 +26,6 @@ import com.aos.AOSBE.Mapper.PriceHistoriesMapper;
 import com.aos.AOSBE.Mapper.ProductItemsMapper;
 import com.aos.AOSBE.Mapper.PromotionProductsMapper;
 import com.aos.AOSBE.Mapper.PromotionsMapper;
-import com.aos.AOSBE.Repository.CostHistoriesRepository;
-import com.aos.AOSBE.Repository.OrderItemsRepository;
-import com.aos.AOSBE.Repository.PriceHistoriesRepository;
-import com.aos.AOSBE.Repository.ProductItemsRepository;
-import com.aos.AOSBE.Repository.PromotionProductsRepository;
-import com.aos.AOSBE.Repository.PromotionsRepository;
-import com.aos.AOSBE.Repository.ReturnsRepository;
-import com.aos.AOSBE.Repository.ReviewsRepository;
 
 @Service
 public class ProductItemsService {
@@ -83,6 +77,7 @@ public class ProductItemsService {
 		Specification<ProductItems> spec = specBuilder.buildFilter(filters);
 		return productItemsRepository.findAll(spec, pageable);
 	}
+
 
 	public List<ProductItems> productItemsFindByBaseProductId(int id) {
 		return productItemsRepository.findByBaseProductsId(id);
@@ -147,6 +142,9 @@ public class ProductItemsService {
 
 	@Transactional
 	public void productItemsDeleteById(int id) {
+		if (!orderItemsRepository.findAllByProductItemId(id).isEmpty()) {
+			throw new RuntimeException("Sản phẩm đã được mua, không thể xóa.");
+		}
 		productItemsRepository.deleteById(id);
 	}
